+++ lib/locatine.rb
@@ -1,2 +1,3 @@
 require 'locatine/search'
+require 'locatine/scope'
 require 'locatine/version'